Update index.js
I used the crawler described here: https://github.com/SAP/project-portal-for-innersource/pull/18 and it created a repos.json file with the topics key:value pair inside of the _InnerSourceMetadata key. Hence to get the portal to work such that all tags were shown, I had to change line 316 from `"topics":oRepo.topics,` to `"topics": oRepo._InnerSourceMetadata.topics,`. Once I did that, topics appeared in the portal. 

I suspect there may be a change in the crawler not reflected in the portal demo?